requirements/base.txt
alabaster==0.7.13
amqp==5.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arviz==0.19.0
asgiref>=3.8.1
astroid==2.15.6
asttokens==2.4.1
async-timeout==4.0.2
asyncpg==0.28.0
attrs==23.2.0
Babel==2.15.0
backcall==0.2.0
bcrypt==4.0.1
beat==2.0.2
billiard==4.2.0
black==23.7.0
cachetools==5.5.0
celery==5.4.0
certifi==2023.7.22
cffi==1.15.1
cfgv==3.3.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==3.0.0
colorama==0.4.6
cons==0.4.6
contourpy==1.2.1
coverage==7.2.7
cryptography==41.0.2
cycler==0.12.1
decorator==5.1.1
dill==0.3.7
distlib==0.3.7
Django>=5.1.1
django-allauth
django-appconf==1.0.6
django-autocomplete-light==3.11.0
django-bootstrap5==24.2
django-cleanup
django-cors-headers==4.2.0
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
django-extensions==3.2.3
django-guardian
django-htmx
django-redis==5.3.0
django-select2==8.2.0
django-stubs==4.2.3
django-stubs-ext==5.0.2
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-gis==1.0
dm-tree==0.1.8
docutils==0.20.1
drf-spectacular==0.26.5
es-core-news-sm 
et-xmlfile==1.1.0
etuples==0.3.9
exceptiongroup==1.1.2
executing==1.2.0
factory-boy==3.3.0
Faker==19.2.0
filelock>=3.15
flake8==6.0.0
flake8-isort==6.0.0
fonttools==4.53.0
gunicorn==21.2.0
h5netcdf==1.3.0
h5py==3.11.0
hiredis==2.2.3
identify==2.5.26
idna==3.4
imagesize==1.4.1
inflection==0.5.1
iniconfig==2.0.0
ipython==8.14.0
iso8601==2.0.0
isort==5.12.0
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kombu==5.4.1
lazy-object-proxy==1.9.0
livereload==2.6.3
logical-unification==0.4.6
lxml==4.6.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mccabe==0.7.0
mdurl==0.1.2
miniKanren==1.0.3
multipledispatch==1.0.0
mypy==1.4.1
mypy-extensions==1.0.0
nltk==3.9.1
nodeenv==1.8.0
numpy==1.25.2
openpyxl==3.1.3
osmiter==1.3.1
packaging==23.1
pandas==2.2.2
paramiko==3.2.0
parso==0.8.3
pathspec==0.11.1
pdfkit
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.10.0
plotly==5.22.0
pluggy==1.2.0
pre-commit==3.3.3
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==6.0.0
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.15.1
pylance==0.6.2
pylint==2.17.5
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pymc==5.16.2
PyNaCl==1.5.0
pypng==0.20220715.0
pyrocko==2024.1.10
pyparsing==3.1.2
pyroutelib3==1.7.2
pytest==7.4.0
pytest-django==4.5.2
pytest-sugar==0.9.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
redis==4.6.0
referencing==0.35.1
regex==2024.11.6
reportlab==4.2.2
requests==2.31.0
requests-file==2.1.0
rich==13.8.1
rpds-py==0.18.1
scikit-learn
scipy==1.14.1
sentry-sdk==1.29.2
setuptools==69.5.1
shapely==2.0.4
six==1.16.0
snowballstemmer==2.2.0
spacy
spacy-legacy
spacy-loggers
Sphinx==7.1.0
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.4
sshtunnel==0.4.0
stack-data==0.6.2
tenacity==8.3.0
termcolor==2.3.0
text-unidecode==1.3
threadpoolctl==3.5.0
tldextract==2.2.3
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.1
tornado==6.3.2
tqdm==4.66.5
traitlets==5.9.0
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.11
typing_extensions==4.7.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.0.4
virtualenv==20.24.2
wcwidth==0.2.6
whitenoise==6.5.0
beautifulsoup4==4.10.0
unidecode
scikit-learn
xhtml2pdf>=0.2.17
django-ckeditor
django-celery-beat


